pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keel"
version = "0.0.0"
description = "investing strategies simulator"
requires-python = ">=3.8"
dependencies = [
    "pandas-datareader>=0.10.0",
    "requests>=2.32.4",
    "typer",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
keel = "keel.cli:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
